Move PioneerEntityBase to .entity_base to avoid circular import
