+++ custom_components/pioneer_async/entity_base.py
@@ -1,0 +1,53 @@
+"""Pioneer AVR entity base."""
+
+import logging
+
+from aiopioneer import PioneerAVR
+from aiopioneer.const import Zones
+
+from homeassistant.helpers.device_registry import DeviceInfo
+from homeassistant.helpers.entity import Entity
+from homeassistant.util import slugify
+
+from .debug import Debug
+
+_LOGGER = logging.getLogger(__name__)
+
+
+def _debug_atlevel(level: int, category: str = __name__):
+    return Debug.atlevel(None, level, category)
+
+
+class PioneerEntityBase(Entity):
+    """Pioneer AVR base entity class."""
+
+    _attr_should_poll = False
+    _attr_has_entity_name = True
+
+    def __init__(
+        self,
+        pioneer: PioneerAVR,
+        device_info: DeviceInfo,
+        zone: Zones | None = None,
+    ) -> None:
+        """Initialize the Pioneer AVR display sensor."""
+        if _debug_atlevel(9):
+            _LOGGER.debug("%s.__init__()", type(self).__name__)
+        self.pioneer = pioneer
+        self.zone = zone
+        self._attr_device_info = device_info
+
+    @property
+    def unique_id(self) -> str:
+        """Return the unique id."""
+        entry_id = self.platform.config_entry.entry_id
+        name = self._attr_name or "device_entity"
+        zone_suffix = "-" + str(self.zone) if self.zone is not None else ""
+        return f"{entry_id}{zone_suffix}-{slugify(name)}"
+
+    @property
+    def available(self) -> bool:
+        """Returns whether the device is available."""
+        return self.pioneer.available and (
+            self.zone is None or self.zone in self.pioneer.zones
+        )
