+++ custom_components/pioneer_async/sensor.py
@@ -22,18 +22,16 @@ from homeassistant.helpers.entity import DeviceInfo
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from . import PioneerEntityBase
 from . import select_dict, reject_dict
-
 from .const import (
     DOMAIN,
     ATTR_PIONEER,
     ATTR_COORDINATORS,
     ATTR_DEVICE_INFO,
 )
-
 from .coordinator import PioneerAVRZoneCoordinator
 from .debug import Debug
+from .entity_base import PioneerEntityBase
 
 
 _LOGGER = logging.getLogger(__name__)
